HTML Links
More info on the difference b/w parent, self, and top in target value.
Interesting to change the default colors by using styles following:
a.link/a.visited/a.hover/a.active. Img file doesn't show up on the web
page and tried figuring out, but I didn't get the answer for the problem
yet tonight.